move saving to utils
